rptest: move AdvertisedTierConfig, AdvertisedTierConfigs, CloudTierName into redpanda_cloud.py
