+++ tests/rptest/services/redpanda.py
@@ -54,10 +54,9 @@ from rptest.clients.rp_storage_tool import RpStorageTool
 from rptest.services import tls
 from rptest.services.admin import Admin
 from rptest.services.redpanda_installer import RedpandaInstaller, VERSION_RE as RI_VERSION_RE, int_tuple as ri_int_tuple
-from rptest.services.redpanda_cloud import CloudCluster, load_tier_profiles, tiers_config_filename
+from rptest.services.redpanda_cloud import CloudCluster, load_tier_profiles, tiers_config_filename, AdvertisedTierConfig, AdvertisedTierConfigs, CloudTierName
 from rptest.services.rolling_restarter import RollingRestarter
 from rptest.services.storage import ClusterStorage, NodeStorage, NodeCacheStorage
-from rptest.services.storage_tiers import AdvertisedTierConfig, AdvertisedTierConfigs, CloudTierName
 from rptest.services.storage_failure_injection import FailureInjectionConfig
 from rptest.services.utils import BadLogLines, NodeCrash
 from rptest.util import inject_remote_script, ssh_output_stderr, wait_until_result
